Fix some eslint issues. There are more.
